Extracted compoents from AttachableEditor
