+++ src/js/shot-generator-new/components/AttachableEditor/ItemSettings.js
@@ -1,0 +1,11 @@
+
+const GUTTER_SIZE = 5
+const ITEM_WIDTH = 68
+const ITEM_HEIGHT = 132
+
+const IMAGE_WIDTH = ITEM_WIDTH
+const IMAGE_HEIGHT = 100
+
+const NUM_COLS = 4
+
+export { GUTTER_SIZE, ITEM_WIDTH, ITEM_HEIGHT, IMAGE_WIDTH, IMAGE_HEIGHT, NUM_COLS }
